Move Integration Tests into Sub-Packages
