+++ integrated-tests/src/test/scala/integration/facia/ShowMoreTest.scala
@@ -1,5 +1,6 @@
-package integration
+package integration.facia
 
+import integration.SharedWebDriver
 import org.scalatest.tags.Retryable
 import org.scalatest.{DoNotDiscover, FlatSpec, Matchers}
